requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
PyAudio==0.2.14
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Wave==0.0.2
Werkzeug==3.1.3
